+++ docs/changelog.md
@@ -14,6 +14,7 @@ Click on the version in the following list to see all changes since the previous
 
 * feat: warning for invalid subtask priority (#62)
 * refactor: simplify code in subtask creation service
+* docs: improve documentation for custom fields
 
 ### [23.02.0] - 2023-02-05