+++ docs/changelog.md
@@ -15,6 +15,7 @@ Click on the version in the following list to see all changes since the previous
 * feat: warning for invalid subtask priority (#62)
 * refactor: simplify code in subtask creation service
 * docs: improve documentation for custom fields
+* feat: extend syntax and allow to set fix version and affected version (#56)
 
 ### [23.02.0] - 2023-02-05