+++ docs/changelog.md
@@ -12,6 +12,12 @@ Click on the version in the following list to see all changes since the previous
 
 ### [Unreleased]
 
+* ...
+
+### [23.02.1] - 2023-02-19
+
+**Syntax now features due date and version fields**
+
 * feat: warning for invalid subtask priority (#62)
 * refactor: simplify code in subtask creation service
 * docs: improve documentation for custom fields
@@ -154,7 +160,8 @@ Click on the version in the following list to see all changes since the previous
 * feat: check for valid license in subtask dialog (#6)
 * feat: restrict the maximum length of a template to 4000 characters
 
-[Unreleased]: https://github.com/codescape/jira-multiple-subtasks/compare/23.02.0...HEAD
+[Unreleased]: https://github.com/codescape/jira-multiple-subtasks/compare/23.02.1...HEAD
+[23.02.1]: https://github.com/codescape/jira-multiple-subtasks/compare/23.02.0...23.02.1
 [23.02.0]: https://github.com/codescape/jira-multiple-subtasks/compare/23.01.0...23.02.0
 [23.01.0]: https://github.com/codescape/jira-multiple-subtasks/compare/22.12.0...23.01.0
 [22.12.0]: https://github.com/codescape/jira-multiple-subtasks/compare/22.11.1...22.12.0